+++ cmd/neofs-node/object.go
@@ -213,8 +213,6 @@ func initObjectService(c *cfg) {
 		policer.WithHeadTimeout(c.appCfg.Policer.HeadTimeout),
 		policer.WithReplicator(c.replicator),
 		policer.WithMaxCapacity(c.appCfg.Policer.MaxWorkers),
-		policer.WithPool(c.cfgObject.pool.replication),
-		policer.WithNodeLoader(c),
 		policer.WithNetwork(c),
 		policer.WithReplicationCooldown(c.appCfg.Policer.ReplicationCooldown),
 		policer.WithObjectBatchSize(c.appCfg.Policer.ObjectBatchSize),